refactor: index page style
