+++ src/components/Navbar.js
@@ -5,6 +5,7 @@ import PropTypes from 'prop-types';
 import { useRouter } from 'next/router';
 import { setUser } from 'reducers/users';
 import { getUserProfile } from 'axios/User';
+import styles from 'styles/components/navbar.module.scss';
 import LoginModal from './Login';
 
 const logo = 'images/containers/Navbar/devkor_logo.svg';
@@ -66,9 +67,9 @@ export const DesktopNavbar = () => {
         }}
         title="LOGIN/SIGNUP"
       />
-      <div className="navbar__container">
+      <div className={styles.navbar__container}>
         <div
-          className="logo"
+          className={styles.logo}
           onClick={() => {
             moveTo('/');
           }}
@@ -113,7 +114,7 @@ export const DesktopNavbar = () => {
           />
         </ul>
       </div>
-      <div className="emptyblock" />
+      <div className={styles.emptyblock} />
     </>
   );
 };
@@ -123,17 +124,17 @@ export const Navbar = () => {
   if (isMobile) {
     return (
       <>
-        <div className="navbar__container">
-          <div className="logo">
+        <div className={styles.navbar__container}>
+          <div className={styles.logo}>
             <img src={logo} alt="devkor" />
           </div>
-          <div className="navbar__hamburgermenu">
+          <div className={styles.navbar__hamburgermenu}>
             <div />
             <div />
             <div />
           </div>
         </div>
-        <div className="emptyblock" />
+        <div className={styles.emptyblock} />
       </>
     );
   }
