+++ src/containers/Home/Banner1.js
@@ -1,7 +1,8 @@
 import React from 'react';
+import styles from 'styles/containers/banner1.module.scss';
 
 export const Banner1 = () => {
-  return <div className="home__banner1">대충 배너 1 무언가</div>;
+  return <div className={styles.home__banner1}>대충 배너 1 무언가</div>;
 };
 
 export default Banner1;
