+++ src/containers/Home/Banner2.js
@@ -1,28 +1,29 @@
 import React from 'react';
+import styles from 'styles/containers/banner2.module.scss';
 
 export const Banner2 = () => {
   return (
-    <div className="home__banner2">
+    <div className={styles.home__banner2}>
       <span>TAGS</span>
-      <div className="home__banner2__tagbox">
-        <div className="home__banner2__tagbox__item">Academic</div>
-        <div className="home__banner2__tagbox__item">Academic</div>
-        <div className="home__banner2__tagbox__item">Academic</div>
-        <div className="home__banner2__tagbox__item">Academic</div>
+      <div className={styles.home__banner2__tagbox}>
+        <div className={styles.home__banner2__tagbox__item}>Academic</div>
+        <div className={styles.home__banner2__tagbox__item}>Academic</div>
+        <div className={styles.home__banner2__tagbox__item}>Academic</div>
+        <div className={styles.home__banner2__tagbox__item}>Academic</div>
       </div>
-      <div className="home__banner2__tagbox">
-        <div className="home__banner2__tagbox__item">Academic</div>
-        <div className="home__banner2__tagbox__item">Academic</div>
-        <div className="home__banner2__tagbox__item">Academic</div>
-        <div className="home__banner2__tagbox__item">Academic</div>
-        <div className="home__banner2__tagbox__item">Academic</div>
+      <div className={styles.home__banner2__tagbox}>
+        <div className={styles.home__banner2__tagbox__item}>Academic</div>
+        <div className={styles.home__banner2__tagbox__item}>Academic</div>
+        <div className={styles.home__banner2__tagbox__item}>Academic</div>
+        <div className={styles.home__banner2__tagbox__item}>Academic</div>
+        <div className={styles.home__banner2__tagbox__item}>Academic</div>
       </div>
-      <div className="home__banner2__tagbox">
-        <div className="home__banner2__tagbox__item">Academic</div>
-        <div className="home__banner2__tagbox__item">Academic</div>
-        <div className="home__banner2__tagbox__item">Academic</div>
-        <div className="home__banner2__tagbox__item">Academic</div>
-        <div className="home__banner2__tagbox__item">Academic</div>
+      <div className={styles.home__banner2__tagbox}>
+        <div className={styles.home__banner2__tagbox__item}>Academic</div>
+        <div className={styles.home__banner2__tagbox__item}>Academic</div>
+        <div className={styles.home__banner2__tagbox__item}>Academic</div>
+        <div className={styles.home__banner2__tagbox__item}>Academic</div>
+        <div className={styles.home__banner2__tagbox__item}>Academic</div>
       </div>
     </div>
   );
