+++ src/styles/components/auth.module.scss
@@ -1,14 +1,5 @@
 // components - login.js
 
-.background {
-  position: fixed;
-  top: 0px;
-  left: 0px;
-  width: 100vw;
-  height: 100vh;
-  background-color: rgba(102, 102, 102, 0.5);
-  z-index: 5;
-}
 .auth {
   position: fixed;
   top: 50%;
