+++ src/styles/components/modal.module.scss
@@ -1,0 +1,9 @@
+.background {
+  position: fixed;
+  top: 0px;
+  left: 0px;
+  width: 100vw;
+  height: 100vh;
+  background-color: rgba(102, 102, 102, 0.5);
+  z-index: 5;
+}
