+++ src/styles/components/navbar.module.scss
@@ -1,0 +1,69 @@
+.emptyblock {
+  height: calc(4.2vh + 36px);
+  visibility: hidden;
+}
+
+.navbar__container {
+  width: 85.4%;
+  height: 4.2vh;
+  padding-left: 7.3%;
+  padding-right: 7.3%;
+  padding-top: 17px;
+  padding-bottom: 19px;
+  justify-content: space-between;
+  align-items: center;
+  display: inline-flex;
+  position: fixed;
+  z-index: 3;
+  background-color: white;
+  top: 0;
+  .logo {
+    width: 100px;
+    height: 50px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    color: white;
+    > img {
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+    }
+  }
+  > ul {
+    list-style: none;
+    display: inline-flex;
+    margin: 0;
+    padding: 0;
+    li {
+      margin-right: 30px;
+      height: 50px;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      font-weight: normal;
+      font-stretch: normal;
+      font-style: normal;
+      line-height: 1.34;
+      letter-spacing: normal;
+      color: #000000;
+    }
+    li:nth-last-child(1) {
+      margin-right: 0px;
+    }
+  }
+  .navbar__hamburgermenu {
+    width: 25px;
+    height: 40px;
+    display: inline-flex;
+    flex-direction: row;
+    justify-content: space-between;
+    transform: rotate(90deg);
+    div {
+      width: 3px;
+      height: 100%;
+      border-radius: 10px;
+      background-color: black;
+    }
+  }
+}
