+++ src/styles/containers/banner1.module.scss
@@ -1,0 +1,12 @@
+.home__banner1 {
+  width: 100vw;
+  position: absolute;
+  left: 0;
+  height: 33.1vh;
+  background-color: black;
+  color: white;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 26px;
+}
