+++ src/styles/containers/banner2.module.scss
@@ -1,0 +1,49 @@
+.home__banner2 {
+  width: 100%;
+  height: auto;
+  padding-top: 33.1vh;
+  color: black;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  margin-bottom: 75px;
+  > span {
+    font-family: OpenSans;
+    margin-bottom: 20px;
+    font-size: 32px;
+    font-weight: normal;
+    font-stretch: normal;
+    font-style: normal;
+    line-height: 1.34;
+    letter-spacing: normal;
+    text-align: center;
+    color: #000000;
+    margin-top: 40px;
+    -webkit-text-stroke: 1px #707070;
+  }
+  .home__banner2__tagbox {
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    margin-bottom: 8px;
+    .home__banner2__tagbox__item {
+      width: 130px;
+      height: 40px;
+      border-radius: 31px;
+      border: solid 1px #707070;
+      background-color: #ffffff;
+      margin-right: 16px;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      @media (min-width: 0px) and (max-width: 767.99px) {
+        font-size: 0.5vw;
+        width: 15%;
+        box-sizing: border-box;
+      }
+    }
+    div:nth-last-child(1) {
+      margin-right: 0px;
+    }
+  }
+}
