+++ src/styles/containers/banner3.module.scss
@@ -1,0 +1,18 @@
+.home__banner3 {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  .home__banner3__projectwrapper {
+    display: inline-flex;
+    margin-top: 12px;
+    div {
+      margin-right: 18px;
+    }
+    div:nth-last-child(1) {
+      margin-right: 0px;
+    }
+  }
+  :first-child {
+    margin-top: 0px;
+  }
+}
