+++ src/styles/index.module.scss
@@ -1,0 +1,9 @@
+.wrapper {
+  padding-left: 7.3%;
+  padding-right: 7.3%;
+  width: 85.4%;
+  position: relative;
+}
+.home__container {
+  width: 100%;
+}
